+++ src/io/blockchain_request_handlers.h
@@ -1,0 +1,107 @@
+/*
+ * blockchain_request_handlers.h
+ *
+ * @copywrite  2018 Devvio Inc
+ */
+#pragma once
+
+#include <boost/filesystem.hpp>
+
+#include "pbuf/devv_pbuf.h"
+#include "primitives/blockchain.h"
+#include "primitives/block_tools.h"
+
+namespace Devv {
+
+/**
+ * Read a persisted block
+ *
+ * @param chain
+ * @param shard_name
+ * @param block
+ * @param working_dir
+ * @return
+ */
+std::vector<byte> ReadBlock(const Blockchain& chain,
+                            const std::string& shard_name,
+                            size_t block,
+                            const boost::filesystem::path& working_dir);
+
+std::vector<byte> ReadBlock(const boost::filesystem::path& shard_path, size_t segment_index, size_t segment_height);
+
+std::string GetServiceArgument(std::map<std::string, std::string>& args, const std::string& key);
+
+bool HasShard(const std::string& shard, const std::string& working_dir);
+
+bool HasBlock(const Blockchain& chain,
+              const std::string& shard_name,
+              size_t block,
+              const boost::filesystem::path& working_dir);
+
+bool HasBlock(const boost::filesystem::path& shard_path, size_t segment_index, size_t segment_height);
+
+std::map<std::string, std::string> TraceTransactions(const std::string& shard_name,
+                                                     uint32_t start_block,
+                                                     uint32_t end_block,
+                                                     const std::vector<byte>& target,
+                                                     const boost::filesystem::path& working_dir,
+                                                     const Blockchain& chain);
+
+/**
+ * Dispatches incoming query requests.
+ */
+class ServiceRequestEventHandler {
+ public:
+  /**
+   * Consturctor
+   * @param working_dir
+   * @param shard_name
+   */
+  ServiceRequestEventHandler(const boost::filesystem::path& working_dir,
+                             const std::string& shard_name);
+
+  ServiceResponsePtr dispatchRequest(const Blockchain& chain,
+                                     const ServiceRequest& request,
+                                     const ServiceResponse& response);
+
+  ServiceResponsePtr handleBlockInfoRequest(const Blockchain& chain,
+                                            ServiceRequestPtr request,
+                                            ServiceResponsePtr response);
+
+  ServiceResponsePtr handleShardInfoRequest(const Blockchain& chain,
+                                            ServiceRequestPtr request,
+                                            ServiceResponsePtr response);
+
+  ServiceResponsePtr handleTraceRequest(const Blockchain& chain,
+                                        ServiceRequestPtr request,
+                                        ServiceResponsePtr response);
+
+  ServiceResponsePtr handleTxInfoRequest(const Blockchain& chain,
+                                         ServiceRequestPtr request,
+                                         ServiceResponsePtr response);
+
+ private:
+  const boost::filesystem::path working_dir_;
+
+  const std::string shard_name_;
+};
+
+/**
+ * Handle service request operations.
+ * See the Repeater Interface page of the wiki for further information.
+ * @param request - a smart pointer to the request
+ * @param working_dir - the directory where blocks are stored
+ * @param shard_name - the name of the shard this process tracks
+ * @param shard_index - the index of the shard this process tracks
+ * @param chain - the blockchain this shard tracks
+ * @return ServiceResponsePtr - a smart pointer to the response
+ */
+ServiceResponsePtr HandleServiceRequest(ServiceRequestEventHandler& event_handler,
+                                        const ServiceRequestPtr request,
+                                        const std::string& working_dir,
+                                        const std::string& shard_name,
+                                        unsigned int shard_index,
+                                        const Blockchain& chain);
+
+
+} // namespace Devv
